some updates

+++ src/app/todo/todo.component.ts
@@ -1,6 +1,6 @@
 import { moveItemInArray } from '@angular/cdk/drag-drop';
 import { Component, ElementRef, OnInit, TemplateRef, ViewChild } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { AuthService } from '../auth/auth.service';
 import { Category } from '../models/category';
@@ -22,6 +22,7 @@ export class TodoComponent implements OnInit {
   categorisedTasks: {[key: string] : Task[]} = {};
   startDate: Date;
   task: Task;
+  dialogRef: MatDialogRef<any, any>;
 
   constructor(
     private taskService: TaskService,
@@ -38,6 +39,7 @@ export class TodoComponent implements OnInit {
           return -1;
         }
       });
+      console.log(this.categories);
       this.taskService.getTasks().subscribe((resp: any) => {
         if(resp.length) {
           this.tasks = resp;
@@ -58,9 +60,9 @@ export class TodoComponent implements OnInit {
   }
 
   openDialogWithTemplateRef(templateRef: TemplateRef<any>) {
-    const dialogRef = this.dialog.open(templateRef);
+    this.dialogRef = this.dialog.open(templateRef);
 
-    dialogRef.afterClosed().subscribe(() => {
+    this.dialogRef.afterClosed().subscribe(() => {
       this.taskService.addTask(this.task).subscribe((resp: any) => {
         if(resp.taskCreation) {
           console.log(resp);
@@ -102,12 +104,20 @@ export class TodoComponent implements OnInit {
     return categorisedTasks;
   }
 
-  onAddTask(categoryName: string, taskName: string, dueDate: Date) {
+  // Es existiert noch ein Bug wenn categoryName not null aber category nicht existiert
+  async onAddTask(categoryName: string, taskName: string, dueDate: Date) {
     const userID: string = this.authService.getUserID();
-    const category: Category = this.categories.find(category => category.name === categoryName);
-    if(!category) {
+    let category: Category = this.categories.find(category => category.name === categoryName);
+    if(!category && categoryName === "") {
       this.snackBar.open('Bitte die Kategorie erst hinzufügen.', 'OK', { duration: 3000 });
       return
+    } else if (!category && categoryName) {
+      this.onAddCategory(categoryName);
+      category = this.categories.find(category => category.name === categoryName);
+      if(!category) {
+        this.snackBar.open('Es ist zu einem Fehler gekommen', 'OK', {duration: 3000});
+        return;
+      }
     }
     const task : Task = { 
       category: category,
@@ -118,6 +128,7 @@ export class TodoComponent implements OnInit {
       done: false
     }
     this.task = task;
+    this.dialogRef.close();
   }
 
   onDeleteTask(task: Task) {
@@ -135,9 +146,9 @@ export class TodoComponent implements OnInit {
   onAddCategory(categoryName: string) {
     if(categoryName) {
       this.taskService.newCategory(categoryName).subscribe((resp: { message: string, categoryCreation: boolean, newCategory?: Category, reason: string }) => {
-        console.log(resp);
         if(resp.categoryCreation) {
           this.categories.push(resp.newCategory);
+          this.categories.sort((a,b) => a.name > b.name ? 1 : -1);
           this.snackBar.open('Kategorie wurde hinzugefügt', 'OK' , { duration: 3000 });
           this.newCategory.nativeElement.value = '';
         } else if (!resp.categoryCreation) {
@@ -151,9 +162,13 @@ export class TodoComponent implements OnInit {
   }
 
   onDeleteCategory(category: string) {
+    if(Object.keys(this.categorisedTasks).includes(category)){
+      this.snackBar.open(`Bitte lösche zunächst die Aufgaben, die zu der Kategorie ${category} gehören.`, "OK", {duration: 3000});
+      return;
+    }
     this.taskService.deleteCategory(category).subscribe((resp: Category) => {
       console.log(resp);
-      this.categories.splice(this.categories.indexOf({ userID: this.authService.getUserID(), name: category}), 1);
+      this.categories.splice(this.categories.indexOf(resp), 1);
       this.newCategory.nativeElement.value = "";
       this.snackBar.open('Kategorie wurde entfernt', 'OK', { duration: 3000 });
     })
